Fix HomeWork_4_1

+++ HomeWork_4/Example_1/Program.cs
@@ -2,6 +2,6 @@
 // Напишите цикл, который принимает на вход два числа (A и B) и возводит число A в натуральную степень B.
 
 double number = EnteringData.DoubleData("Введите число: ");
-double degree = EnteringData.DoubleData("Введите степень: ");
+int degree = EnteringData.IntData("Введите степень: ");
 
 Console.WriteLine(Calculation.Pow(number,degree));
